+++ packages/hooks/src/index.ts
@@ -1,3 +1,4 @@
+// Export all hooks
 export * from "./useAssert";
 export * from "./useBool";
 export * from "./useBreakpoints";